Move app stuff around to files
