+++ debugger/src/Data.ts
@@ -1,0 +1,5 @@
+export type Id = number;
+
+export interface HasId {
+    id: Id
+}
